requirements.txt
accelerate==0.31.0
bitsandbytes==0.43.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
customtkinter==4.6.2
darkdetect==0.8.0
diffuser==0.0.1
diffusers==0.29.0
filelock==3.15.1
fsspec==2024.6.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==7.1.0
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.3.0
protobuf==5.27.1
psutil==5.9.8
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
sentencepiece==0.2.0
sympy==1.12.1
tbb==2021.12.0
tk==0.1.0
tokenizers==0.19.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
urllib3==2.2.1
zipp==3.19.2
